Move common setUp logic to base test class
